DG22-2765 adjust hvac1 eligibility
