+++ src/pages/residential_ac/ActivityRequirementsResAC.jsx
@@ -1,18 +1,10 @@
 import React, { Fragment, useState, useEffect } from 'react';
 
 import { ProgressIndicator } from 'nsw-ds-react/forms/progress-indicator/progressIndicator';
-import OpenFiscaAPI from 'services/openfisca_api';
 import SpinnerFullscreen from 'components/layout/SpinnerFullscreen';
 import HeroBanner from 'nsw-ds-react/heroBanner/heroBanner';
 import LoadClausesResidentialActivityRequirements from './LoadClausesActivityRequirements';
 import { IS_DRUPAL_PAGES } from 'types/app_variables';
-import { 
-  HVAC1_PDRSAug24_AEER_greater_than_minimum,
-  HVAC1_PDRSAug24_TCPSF_greater_than_minimum,
-  HVAC1_PDRSAug24_HSPF_mixed_eligible,
-  HVAC1_PDRSAug24_HSPF_cold_eligible,
-  HVAC1_PDRSAug24_ACOP_eligible,
-} from 'types/openfisca_variables';
 import { FormGroup, Select } from 'nsw-ds-react/forms';
 import { USER_TYPE_OPTIONS } from 'constant/user-type';
 import {
@@ -26,21 +18,18 @@ import MoreOptionsCard from 'components/more-options-card/more-options-card';
 import { BASE_RESIDENTIAL_AC_ELIGIBILITY_ANALYTICS_DATA } from 'constant/base-analytics-data';
 
 export default function ActivityRequirementsResAC(props) {
-  const { entities, variables, setEntities, setVariables } = props;
+  const { entities, variables, loading } = props;
 
   const [formValues, setFormValues] = useState([]);
   const [stepNumber, setStepNumber] = useState(1);
-  const [dependencies, setDependencies] = useState([]);
   const [variableToLoad, setVariableToLoad] = useState(
     'HVAC1_PDRSAug24_installation_replacement_final_activity_eligibility',
   );
-  const [variable, setVariable] = useState({});
   const [clausesForm, setClausesForm] = useState([]);
   const [showError, setShowError] = useState(false);
   const [userType, setUserType] = useState('');
   const [isUserTypeValid, setIsUserTypeValid] = useState(true);
   const [userTypeError, setUserTypeError] = useState('');
-  const [loading, setLoading] = useState(true);
 
   useEffect(() => {
     window.scrollTo(0, 0);
@@ -48,55 +37,6 @@ export default function ActivityRequirementsResAC(props) {
     updateEstimatorFormAnalytics(BASE_RESIDENTIAL_AC_ELIGIBILITY_ANALYTICS_DATA);
     updateFeedbackFormAnalytics(BASE_RESIDENTIAL_AC_ELIGIBILITY_ANALYTICS_DATA);
   }, []);
-
-  useEffect(() => {
-    OpenFiscaAPI.getVariable(variableToLoad)
-      .then((res) => {
-        setVariable(res.data);
-        setLoading(false);
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  }, []);
-
-  useEffect(() => {
-    if (Object.keys(variable).length && stepNumber === 1) {
-      const children = variable.input_offsprings;
-
-      // Define the original array (at a minimum include the Implementation Date)
-      var array = [];
-
-      var dep_arr = [];
-
-      children.map((child) => {
-        array.push({ ...child, form_value: '', invalid: false, hide: false });
-      });
-
-      array.sort((a, b) => a.metadata.sorting - b.metadata.sorting);
-
-      const names = [
-        HVAC1_PDRSAug24_AEER_greater_than_minimum,
-        HVAC1_PDRSAug24_TCPSF_greater_than_minimum,
-        HVAC1_PDRSAug24_HSPF_mixed_eligible,
-        HVAC1_PDRSAug24_HSPF_cold_eligible,
-        HVAC1_PDRSAug24_ACOP_eligible,
-      ];
-
-      dep_arr = array.filter((item) => names.includes(item.name));
-      array.forEach((item) => {
-        if (names.includes(item.name)) {
-          item.hide = true;
-        }
-      });
-
-      dep_arr = dep_arr.map((obj, i) => ({ ...obj, hide: true }));
-
-      setFormValues(array);
-      setDependencies(dep_arr);
-      setLoading(false);
-    }
-  }, [variable]);
 
   useEffect(() => {
     let new_arr = [];
@@ -219,13 +159,11 @@ export default function ActivityRequirementsResAC(props) {
               )}
               <LoadClausesResidentialActivityRequirements
                 variableToLoad={variableToLoad}
-                variable={variable}
                 variables={variables}
                 entities={entities}
                 stepNumber={stepNumber}
                 setStepNumber={setStepNumber}
                 formValues={formValues}
-                dependencies={dependencies}
                 setFormValues={setFormValues}
                 clausesForm={clausesForm}
                 setClausesForm={setClausesForm}
@@ -235,8 +173,6 @@ export default function ActivityRequirementsResAC(props) {
                   setStepNumber(stepNumber - 1);
                 }}
                 onValidateUserType={onValidateUserType}
-                loading={loading}
-                setLoading={setLoading}
               />
             </>
           )}
